Update main file.
Added check for schedule availability before submitting it. If it is not found, then a message is encountered.
The code will now check if there is a Regular pair to collect a group or a paid transfer, and send a message about if Regular is not found. It's probably better to handle the situation where the requested action is not present on the site or secure data has been entered.